requirements.txt
# Django Backend
Django==6.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
python-dotenv==1.0.0

# Authentication
bcrypt==4.0.1
PyJWT==2.8.0

# Legacy Streamlit (can be removed once React is fully integrated)
streamlit==1.28.0
